Crate page_attribute_table

Support for the Page Attribute Table (PAT) feature on x86.

PAT replaces the legacy Memory Type Range Registers (MTRRs) on x86. PAT allows the system to assign memory types, i.e., specify their caching behavior, to regions of “linear” (virtual) addresses. This is in contrast to MTRRs, which operate on regions of physical addresses.

Structs

PageAttributeTable

The Page Attribute Table (PAT) consists of 8 “slots” that can each be configured with a different MemoryCachingType.

PatNotSupported

An empty error type indicating that the Page Attribute Table is not supported on this machine.

Enums

MemoryCachingType

The various types of memory caching that x86 supports for usage in the PageAttributeTable.

Statics

FIXED_PAT

Theseus’s fixed PageAttributeTable has slots that align with the default meaning of page table entry bits on x86_64.

Functions

init

Sets up and enables the Page Attribute Table (PAT) for this (the current) CPU.

is_supported

Returns true if the Page Attribute Table is supported on this system.